Criação CRUD parte 2

+++ src/main/java/com/academiadigital/Service/AlunoService.java
@@ -1,14 +1,15 @@
 package com.academiadigital.Service;
 
 import com.academiadigital.Entity.Aluno;
+import com.academiadigital.Entity.AvaliacaoFisica;
 import com.academiadigital.Entity.Form.AlunoForm;
 import com.academiadigital.Entity.Form.AlunoUpdateForm;
-import com.academiadigital.Infra.Utils.JavaTimeUtils;
 import com.academiadigital.Repository.AlunoRepository;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.GetMapping;
 
-import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
@@ -36,14 +37,8 @@ public class AlunoService {
         return alunoRepository.findById(id);
     }
 
-    public List<Aluno> findAll(String dataDeNascimento){
-        if(dataDeNascimento == null){
-            return alunoRepository.findAll();
-        } else {
-            LocalDate localDate = LocalDate.parse(dataDeNascimento, JavaTimeUtils.LOCAL_DATE_FORMATTER);
-            return alunoRepository.findByDataDeNascimento(String.valueOf(localDate));
-        }
-
+    public List<Aluno> findAll(){
+        return alunoRepository.findAll();
     }
 
     public Aluno updateAluno(Long id, AlunoUpdateForm updateForm){
@@ -60,7 +55,16 @@ public class AlunoService {
 
     }
 
-    public void deleteAluno(Long id){
-        alunoRepository.deleteById(id);
+    public void deleteAluno(Long id) {
+        if (alunoRepository.existsById(id)){
+            alunoRepository.deleteById(id);
+        }
     }
+
+    @GetMapping
+    public List<AvaliacaoFisica> getAllAvaliacaoFisica(Long id){
+        Aluno aluno = alunoRepository.findById(id).get();
+        return aluno.getAvaliacoes();
+    }
+
 }
